New: Define rule configurations in repository 💖⏫ (#85)
Remove Airbnb dependency and define all rules in repo. Refactor how rulesets are defined and
extended for better grouping of configurations and rules.

BREAKING CHANGE: Rename any configuration using `eloquence/webpack` to `eloquence/react`. This
  better reflects the project type rather than the tools used for the project.

RELEASE NOTES: This release cleans up the dependencies and rule definitions for Eloquence, most
importantly the dependency on the Airbnb package is removed and those rules are
copied into this repo. Reasons for this include:

- Better access to rules to handle switching between dev and test severity
  levels
- Faster releases to rules and linting issues
- More fine grained control of which rules are applied to which lint-set, eg
  Node.js and React rules are only applied to those lint sets.

Going forward this also has the benefit of consolidating all of the lint rules
applied for each set to this repo, allowing this repo to serve as the manager of
ESLint dependencies and rules.